shift getRow functions into memberparse.py
